+++ package.json
@@ -2,7 +2,7 @@
 	"name": "website",
 	"version": "0.1.0",
 	"description": "She Codes PCC website",
-	"main": "pages/index.js",
+	"main": "src/pages/index.js",
 	"author": "She Codes PCC Contributors",
 	"license": "Apache-2.0",
 	"scripts": {
@@ -19,9 +19,12 @@
 	"dependencies": {
 		"next": "12.0.7",
 		"react": "17.0.2",
-		"react-dom": "17.0.2"
+		"react-dom": "17.0.2",
+		"sass": "^1.45.0"
 	},
 	"devDependencies": {
+		"@rmwc/theme": "^6.1.4",
+		"@rmwc/typography": "^6.1.4",
 		"eslint": "8.4.1",
 		"eslint-config-google": "^0.14.0",
 		"eslint-config-next": "12.0.7",
