+++ src/pages/connect-with-us.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import Header from '../components/Header'
 
-export default function ResourcesPage() {
+export default function ConnectWithUsPage() {
 	return (
 		<div>
-			<Header title="Resources" />
+			<Header title="Connect with us 👋" />
 		</div>
 	)
 }
