+++ src/pages/projects.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Header from '../components/Header'
+
+export default function ProjectsPage() {
+	return (
+		<div>
+			<Header title="Projects 🚀" />
+		</div>
+	)
+}
