+++ src/styles/theme.scss
@@ -1,0 +1,9 @@
+// theme
+// more info: https://rmwc.io/styling-theming
+$mdc-theme-primary: #534fc0;
+$mdc-theme-secondary: #b6ffe9;
+$mdc-theme-error: #ba1b1b;
+
+// import RMWC
+@import '@material/theme/dist/mdc.theme.css';
+@import '@rmwc/theme/theme.css';
